requirements.txt
asgiref==3.10.0
beautifulsoup4==4.14.2
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
Django==5.2.7
django-allauth==65.12.1
django-dynamic-breadcrumbs==0.4.3
django-environ==0.12.0
django-redis==6.0.0
django-view-breadcrumbs==2.5.1
environ==1.0
idna==3.11
oauthlib==3.3.1
pillow==12.0.0
psycopg2-binary==2.9.11
pycparser==2.23
PyJWT==2.10.1
redis==7.0.1
requests==2.32.5
soupsieve==2.8
sqlparse==0.5.3
typing_extensions==4.15.0
urllib3==2.5.0
